Componente alterado para Home e generalizado para Comidas e Bebidas
